fix: move speedcurve utils into payment so as not to break module boundaries
