+++ src/payment/cart/OrderSummary.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from '@edx/frontend-i18n';
 
-import markPerformanceIfAble from '../../speedcurve';
+import markPerformanceIfAble from '../speedcurve';
 
 class OrderSummary extends React.Component {
   componentDidMount() {
