+++ src/payment/index.js
@@ -4,3 +4,4 @@ export { default as saga } from './data/sagas';
 export { storeName } from './data/selectors';
 export { default as EcommerceRedirect } from './EcommerceRedirect';
 export { default as responseInterceptor } from './responseInterceptor';
+export { default as markPerformanceIfAble } from './speedcurve';
