+++ src/payment/speedcurve.js
